Add files via upload
This is a text-based dungeon RPG application. The idea of the game is to choose a player class such as paladin, fighter, cleric, etc. that allows the user to enter a room that has a randomly generated monster in it. The player gets to choose an option to either fight, run, exit, check their stats, check the monsters stats, or use a shop. I implemented in game currency (jewels) and also gave each class unique skills and weapons. Felt like a real game programmer which really got the juices flowing!